Authentication, todo owner, passport js, jwt

+++ src/controllers/TodoController.js
@@ -3,15 +3,20 @@ const Todo = require('../models/Todo');
 const todoController = {}
 
 todoController.get = async (request, response, next) => {
+    const { user } = request;
+    
+    const query = {
+        owner: user._id
+    };
+
     try {
-        const todo = await Todo.find();
-        response.status(200);
-        return response.send({
+        const todo = await Todo.find(query).sort({ createdAt: 'desc' });
+
+        return response.status(200).send({
             success: true,
             todo: todo
         });
     } catch (error) {
-        console.log('Error get: Todo Controller with mongoose');
         next(error);
     }
 };
@@ -21,11 +26,13 @@ todoController.getOne = async  (request, response, next) => {
 
     try {
         const todo = await Todo.findById({ _id: id });
-        
-        if(todo === null) return response.status(404).send({
-            success: false,
-            message: 'Todo does not exist in database!'
-        });
+
+        if(todo === null || !todo.owner.equals(request.user._id)) {
+            return response.status(404).send({
+                success: false,
+                message: 'Todo does not exist in database!'
+            });
+        }
 
         return response.status(200).send({
             success: true,
@@ -33,18 +40,16 @@ todoController.getOne = async  (request, response, next) => {
         });
 
     } catch (error) {
-        console.log('Error get one todo: Todo Controller with mongoose');
         next(error);
     }
-}
+};
 
 todoController.create = async (request, response, next) => {
-    console.log('Create Todo');
-
     const { title } = request.body;
     
     const newTodo = new Todo({
-        title: title
+        title: title,
+        owner: request.user
     });
 
     try {
@@ -54,7 +59,6 @@ todoController.create = async (request, response, next) => {
             todo: todo
         });
     } catch (error) {
-        console.log('Error save: Todo Controller with mongoose');
         next(error);
     }
 };
@@ -63,17 +67,16 @@ todoController.update = async (request, response, next) => {
     const id = request.params.id;
     const { title, completed } = request.body;
 
-    // Check if Todo does not exist in the db and retuen false with message ..
-
-    // if(todo === null) return response.status(404).send({
-    //     success: false,
-    //     message: 'Todo does not exist in database!'
-    // });
-
     try {
         const todo = await Todo.updateOne(
-            { _id: id },
-            { title: title, completed: completed }
+            {
+                _id: id,
+                owner: request.user
+            },
+            {
+                title: title,
+                completed: completed
+            }
         );
 
         if(todo.n === 0) {
@@ -85,7 +88,6 @@ todoController.update = async (request, response, next) => {
             todo: todo
         });
     } catch (error) {
-        console.log('Error update: Todo Controller with mongoose');
         next(error);
     }
 }
@@ -94,7 +96,7 @@ todoController.destroy = async (request, response, next) => {
     const id = request.params.id;
 
     try {
-        const result = await Todo.deleteOne({ _id: id });
+        const result = await Todo.deleteOne({ _id: id, owner: request.user });
         if(result.deletedCount === 0) {
             response.status(404);
             return response.send({
@@ -103,11 +105,10 @@ todoController.destroy = async (request, response, next) => {
             });
         }
     } catch (error) {
-        console.log('Error delete: Todo Controller with mongoose');
         next(error);
     }
-    
-    return response.send({
+
+    return response.status(200).send({
         success: true,
         message: 'Todo deleted'
     });
